reorganized directories
